Separate files into packages
