--- src/main/java/AddTaskCommand.java
@@ -1,11 +1,0 @@
-public class AddTaskCommand extends Command {
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
