+++ src/main/java/duke/command/AddTaskCommand.java
@@ -1,7 +1,13 @@
-public class ShowListCommand extends Command {
+package duke.command;
+
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.Ui;
+
+public class AddTaskCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
-        ui.showList(taskList.getTaskList());
+
     }
 
     @Override
