+++ src/main/java/duke/command/DeleteTaskCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.command.Command;
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.Ui;
+
 public class DeleteTaskCommand extends Command {
     private int index;
