+++ src/main/java/duke/command/MarkTaskCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.Ui;
+
 public class MarkTaskCommand extends Command {
     private boolean isDone;
     private int index;
