+++ src/main/java/duke/exception/InvalidCommandException.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class InvalidCommandException extends DukeException {
     public InvalidCommandException(String description) {
         super(description);
