+++ src/main/java/duke/exception/InvalidIndexException.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class InvalidIndexException extends DukeException {
     public InvalidIndexException(String description) {
         super(description);
