+++ src/main/java/duke/exception/MissingDescriptionException.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class MissingDescriptionException extends DukeException {
     public MissingDescriptionException(String description) {
         super(description);
