+++ src/main/java/duke/exception/MissingIndexException.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class MissingIndexException extends DukeException {
     public MissingIndexException(String description) {
         super(description);
