+++ src/main/java/duke/exception/MissingTimeException.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class MissingTimeException extends DukeException {
     public MissingTimeException(String description) {
         super(description);
