+++ src/main/java/duke/main/TaskList.java
@@ -1,3 +1,7 @@
+package duke.main;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
