+++ src/main/java/duke/main/Ui.java
@@ -1,3 +1,7 @@
+package duke.main;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 import java.util.Scanner;
